First tests on server are working

+++ server/src/Food.cpp
@@ -5,7 +5,7 @@
 #include "Food.h"
 #include "Gamefield.h"
 
-Food::Food(GamefieldPtr mGamefield, uint64_t mId, const Vector& mPosition) :
+Food::Food(GamefieldPtr mGamefield, uint32_t mId, const Vector& mPosition) :
 		Element(mGamefield, mId, mPosition, mGamefield->getOptions().food.color, mGamefield->getOptions().food.size,
 				mGamefield->getOptions().food.mass) {
